# column to row refactor for performance