--- a/java/src/main/native/src/row_conversion.cu
+++ b/java/src/main/native/src/row_conversion.cu
@@ -74,8 +74,7 @@ constexpr auto NUM_STRING_ROWS_PER_BLOCK_FROM_ROWS = 64;
 constexpr auto MIN_STRING_BLOCKS = 32;
 constexpr auto MAX_STRING_BLOCKS = MAX_BATCH_SIZE;
 
-constexpr auto NUM_THREADS_IN_WARP = 32;
-constexpr auto NUM_THREADS = 1024;
+constexpr auto NUM_WARPS_IN_BLOCK = 32;
 constexpr auto NUM_VALIDITY_THREADS_PER_TILE = 32;
 
 // needed to suppress warning about cuda::barrier
@@ -595,7 +594,7 @@ __global__ void copy_to_rows(const size_type num_rows, const size_type num_colum
   // any calculation to do here, but it is important to note.
 
   auto group = cooperative_groups::this_thread_block();
-  auto warp = cooperative_groups::tiled_partition<NUM_THREADS_IN_WARP>(group);
+  auto warp = cooperative_groups::tiled_partition<cudf::detail::warp_size>(group);
   extern __shared__ int8_t shared_data[];
 
 #ifdef ASYNC_MEMCPY_SUPPORTED
@@ -625,21 +624,29 @@ __global__ void copy_to_rows(const size_type num_rows, const size_type num_colum
   // by the tiled partition and each thread in that partition works on a row
   for (int relative_col = warp.meta_group_rank(); relative_col < num_tile_cols;
        relative_col += warp.meta_group_size()) {
+
+    auto const absolute_col = relative_col + tile.start_col;
+    auto const col_size = col_sizes[absolute_col];
+    auto const col_offset = col_offsets[absolute_col];
+    auto const relative_col_offset = col_offset - starting_column_offset;
+    auto const col_ptr = input_data[absolute_col];
+
+    if (col_ptr == nullptr) {
+      // variable-width data column
+      continue;
+    }
+
     for (int relative_row = warp.thread_rank(); relative_row < num_tile_rows;
          relative_row += warp.size()) {
-      auto const absolute_col = relative_col + tile.start_col;
 
-      if (input_data[absolute_col] == nullptr || relative_row >= num_tile_rows) {
-        // variable-width data
+      if (relative_row >= num_tile_rows) {
+        // out of bounds
         continue;
       }
       auto const absolute_row = relative_row + tile.start_row;
-      auto const col_size = col_sizes[absolute_col];
-      auto const col_offset = col_offsets[absolute_col];
-      auto const relative_col_offset = col_offset - starting_column_offset;
 
       auto const shared_offset = relative_row * tile_row_size + relative_col_offset;
-      auto const input_src = input_data[absolute_col] + col_size * absolute_row;
+      auto const input_src = col_ptr + col_size * absolute_row;
 
       // copy the element from global memory
       switch (col_size) {
@@ -727,7 +734,7 @@ copy_validity_to_rows(const size_type num_rows, const size_type num_columns,
   // then ballot_sync the bits and write the result to shmem
   // after we fill shared mem memcpy it out in a blob.
   auto group = cooperative_groups::this_thread_block();
-  auto warp = cooperative_groups::tiled_partition<NUM_THREADS_IN_WARP>(group);
+  auto warp = cooperative_groups::tiled_partition<cudf::detail::warp_size>(group);
 
 #ifdef ASYNC_MEMCPY_SUPPORTED
   // Initialize cuda barriers for each tile.
@@ -772,12 +779,12 @@ copy_validity_to_rows(const size_type num_rows, const size_type num_columns,
       // every thread that is participating in the warp has 4 bytes, but it's column-based
       // data and we need it in row-based. So we shuffle the bits around with ballot_sync to
       // make the bytes we actually write.
-      bitmask_type dw_mask = 1;
+      bitmask_type dw_mask = 0x1;
       for (int i = 0; i < threads_per_warp && relative_row + i < num_rows; ++i, dw_mask <<= 1) {
         auto validity_data = __ballot_sync(participation_mask, my_data & dw_mask);
         // lead thread in each warp writes data
         auto const validity_write_offset =
-            validity_data_row_length * (relative_row + i) + relative_col / CHAR_BIT;
+            validity_data_row_length * (relative_row + i) + (relative_col / CHAR_BIT);
         if (warp.thread_rank() == 0) {
           *reinterpret_cast<bitmask_type *>(&shared_data[validity_write_offset]) = validity_data;
         }
@@ -845,7 +852,7 @@ __global__ void copy_strings_to_rows(size_type const num_rows, size_type const n
   // fill out offset/length information for the column. Then all threads of the warp will
   // participate in the memcpy of the string data.
   auto const my_block = cooperative_groups::this_thread_block();
-  auto const warp = cooperative_groups::tiled_partition<NUM_THREADS_IN_WARP>(my_block);
+  auto const warp = cooperative_groups::tiled_partition<cudf::detail::warp_size>(my_block);
 #ifdef ASYNC_MEMCPY_SUPPORTED
   cuda::barrier<cuda::thread_scope_block> block_barrier;
 #endif
@@ -1726,7 +1733,7 @@ void determine_tiles(std::vector<size_type> const &column_sizes,
   auto const square_bias = 32; // bias towards columns for performance reasons
   auto const optimal_square_len = static_cast<size_type>(sqrt(shmem_limit_per_tile));
   auto const desired_tile_height = util::round_up_safe<int>(
-      std::min(optimal_square_len / square_bias, total_number_of_rows), NUM_THREADS_IN_WARP);
+      std::min(optimal_square_len / square_bias, total_number_of_rows), cudf::detail::warp_size);
   auto const tile_height = std::clamp(desired_tile_height, 1, first_row_batch_size);
 
   int row_size = 0;
@@ -1864,29 +1871,26 @@ std::vector<std::unique_ptr<column>> convert_to_rows(
             gpu_batch_row_boundaries, start_col, end_col, tile_height, num_rows, stream);
       });
 
-  // blast through the entire table and convert it
-  dim3 const blocks(gpu_tile_infos.size());
-  dim3 const threads(NUM_THREADS);
-
   // build validity tiles for ALL columns, variable and fixed width.
   auto validity_tile_infos = detail::build_validity_tile_infos(
       tbl.num_columns(), num_rows, shmem_limit_per_tile, batch_info.row_batches);
 
   auto dev_validity_tile_infos = make_device_uvector_async(validity_tile_infos, stream);
-  dim3 const validity_blocks(validity_tile_infos.size());
-  dim3 const validity_threads(NUM_THREADS);
 
   auto const validity_offset = column_info.column_starts.back();
 
-  detail::copy_to_rows<<<blocks, threads, total_shmem_in_bytes, stream.value()>>>(
+  // blast through the entire table and convert it
+  detail::copy_to_rows<<<gpu_tile_infos.size(), NUM_WARPS_IN_BLOCK * cudf::detail::warp_size,
+                         total_shmem_in_bytes, stream.value()>>>(
       num_rows, tbl.num_columns(), shmem_limit_per_tile, gpu_tile_infos, dev_input_data.data(),
       dev_col_sizes.data(), dev_col_starts.data(), offset_functor,
       batch_info.d_batch_row_boundaries.data(),
       reinterpret_cast<int8_t **>(dev_output_data.data()));
 
   // note that validity gets the entire table and not the fixed-width portion
-  detail::copy_validity_to_rows<<<validity_blocks, validity_threads, total_shmem_in_bytes,
-                                  stream.value()>>>(
+  detail::copy_validity_to_rows<<<validity_tile_infos.size(),
+                                  NUM_WARPS_IN_BLOCK * cudf::detail::warp_size,
+                                  total_shmem_in_bytes, stream.value()>>>(
       num_rows, tbl.num_columns(), shmem_limit_per_tile, offset_functor,
       batch_info.d_batch_row_boundaries.data(), dev_output_data.data(), validity_offset,
       dev_validity_tile_infos, dev_input_nm.data());
@@ -1911,7 +1915,6 @@ std::vector<std::unique_ptr<column>> convert_to_rows(
     auto dev_variable_col_output_offsets =
         make_device_uvector_async(column_info.variable_width_column_starts, stream);
 
-    dim3 const string_threads(NUM_THREADS);
     for (uint i = 0; i < batch_info.row_batches.size(); i++) {
       auto const batch_row_offset = batch_info.batch_row_boundaries[i];
       auto const batch_num_rows = batch_info.row_batches[i].row_count;
@@ -1920,7 +1923,8 @@ std::vector<std::unique_ptr<column>> convert_to_rows(
           MAX_STRING_BLOCKS,
           util::div_rounding_up_unsafe(batch_num_rows, NUM_STRING_ROWS_PER_BLOCK_TO_ROWS)));
 
-      detail::copy_strings_to_rows<<<string_blocks, string_threads, 0, stream.value()>>>(
+      detail::copy_strings_to_rows<<<string_blocks, NUM_WARPS_IN_BLOCK * cudf::detail::warp_size, 0,
+                                     stream.value()>>>(
           batch_num_rows, variable_width_table.num_columns(), dev_variable_input_data.data(),
           dev_variable_col_output_offsets.data(), variable_width_offsets->data(),
           column_info.size_per_row, offset_functor, batch_row_offset,
@@ -2223,7 +2227,6 @@ std::unique_ptr<table> convert_from_rows(lists_column_view const &input,
 
   dim3 const blocks(
       util::div_rounding_up_unsafe(gpu_tile_infos.size(), NUM_TILES_PER_KERNEL_FROM_ROWS));
-  dim3 const threads(NUM_THREADS);
 
   // validity needs to be calculated based on the actual number of final table columns
   auto validity_tile_infos =
@@ -2240,7 +2243,8 @@ std::unique_ptr<table> convert_from_rows(lists_column_view const &input,
   if (dev_string_row_offsets.size() == 0) {
     detail::fixed_width_row_offset_functor offset_functor(size_per_row);
 
-    detail::copy_from_rows<<<blocks, threads, total_shmem_in_bytes, stream.value()>>>(
+    detail::copy_from_rows<<<blocks, NUM_WARPS_IN_BLOCK * cudf::detail::warp_size,
+                             total_shmem_in_bytes, stream.value()>>>(
         num_rows, num_columns, shmem_limit_per_tile, offset_functor,
         gpu_batch_row_boundaries.data(), dev_output_data.data(), dev_col_sizes.data(),
         dev_col_starts.data(), gpu_tile_infos, child.data<int8_t>());
@@ -2253,7 +2257,8 @@ std::unique_ptr<table> convert_from_rows(lists_column_view const &input,
 
   } else {
     detail::string_row_offset_functor offset_functor(device_span<size_type const>{input.offsets()});
-    detail::copy_from_rows<<<blocks, threads, total_shmem_in_bytes, stream.value()>>>(
+    detail::copy_from_rows<<<blocks, NUM_WARPS_IN_BLOCK * cudf::detail::warp_size,
+                             total_shmem_in_bytes, stream.value()>>>(
         num_rows, num_columns, shmem_limit_per_tile, offset_functor,
         gpu_batch_row_boundaries.data(), dev_output_data.data(), dev_col_sizes.data(),
         dev_col_starts.data(), gpu_tile_infos, child.data<int8_t>());
@@ -2292,9 +2297,9 @@ std::unique_ptr<table> convert_from_rows(lists_column_view const &input,
     dim3 const string_blocks(
         std::min(std::max(MIN_STRING_BLOCKS, num_rows / NUM_STRING_ROWS_PER_BLOCK_FROM_ROWS),
                  MAX_STRING_BLOCKS));
-    dim3 const string_threads(NUM_THREADS);
 
-    detail::copy_strings_from_rows<<<string_blocks, string_threads, 0, stream.value()>>>(
+    detail::copy_strings_from_rows<<<string_blocks, NUM_WARPS_IN_BLOCK * cudf::detail::warp_size, 0,
+                                     stream.value()>>>(
         offset_functor, dev_string_row_offsets.data(), dev_string_lengths.data(),
         dev_string_col_offsets.data(), dev_string_data_cols.data(), child.data<int8_t>(), num_rows,
         static_cast<cudf::size_type>(string_col_offsets.size()));