Spent some time investigating performance on column to row and row to column conversion code. Lots of wins here. Split PR up, so this is column to row work only. Row to column to follow. Also, planning to add tests and benchmarks into `spark-rapids-jni` once this is all updated.

Starting performance:
```
--------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                      Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------
RowConversion/old_to_row_conversion/1048576/manual_time                     5.80 ms         5.81 ms          116 bytes_per_second=123.315G/s
RowConversion/new_to_row_conversion/1048576/manual_time                     56.8 ms         56.8 ms           12 bytes_per_second=12.5851G/s
RowConversion/new_to_row_extended_conversion/1048576/manual_time            56.8 ms         56.8 ms           12 bytes_per_second=12.5846G/s
RowConversion/string_to_row_extended_conversion/1048576/manual_time         63.6 ms         63.6 ms           11 bytes_per_second=11.0478G/s
RowConversion/old_from_row_conversion/1048576/manual_time                   4.16 ms         4.21 ms          168 bytes_per_second=171.829G/s
RowConversion/new_from_row_conversion/1048576/manual_time                   83.1 ms         83.9 ms            8 bytes_per_second=8.5998G/s
RowConversion/new_from_row_extended_conversion/1048576/manual_time          83.1 ms         83.8 ms            8 bytes_per_second=8.60105G/s
RowConversion/string_from_row_extended_conversion/1048576/manual_time       80.0 ms         86.9 ms            9 bytes_per_second=8.78838G/s
```

Final performance:
```
--------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                      Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------
RowConversion/old_to_row_conversion/1048576/manual_time                     5.32 ms         5.31 ms          123 bytes_per_second=134.364G/s
RowConversion/new_to_row_conversion/1048576/manual_time                     6.90 ms         6.92 ms          101 bytes_per_second=103.563G/s
RowConversion/new_to_row_extended_conversion/1048576/manual_time            6.91 ms         6.93 ms          102 bytes_per_second=103.496G/s
RowConversion/string_to_row_extended_conversion/1048576/manual_time         11.6 ms         11.5 ms           61 bytes_per_second=60.832G/s
RowConversion/old_from_row_conversion/1048576/manual_time                   3.90 ms         3.94 ms          180 bytes_per_second=183.478G/s
RowConversion/new_from_row_conversion/1048576/manual_time                   67.7 ms         68.1 ms           11 bytes_per_second=10.5667G/s
RowConversion/new_from_row_extended_conversion/1048576/manual_time          68.0 ms         68.5 ms           10 bytes_per_second=10.5116G/s
RowConversion/string_from_row_extended_conversion/1048576/manual_time       70.8 ms         71.9 ms           10 bytes_per_second=9.92482G/s
```