@ttnghia 2022-06-07
It is recommended to use `compare.py` to show the benchmark changes. Here is what I do:
```
/usr/bin/python3 cpp/build/_deps/benchmark-src/tools/compare.py benchmarks ./0 ./1
```
Where `./0` and `./1` are the binaries of different test builds. cudf must be built into static lib (`BUILD_SHARE_LIB=OFF`).