@hyperbolic2346 2022-06-07
Updated original benchmarks to show differences. The downside on the row to column is due to changes related to this that didn't play nicely with the other side. The other side is the next PR and will fix that though.