@ttnghia 2022-06-15
One more question: Why this is still here, not in rapids-spark-jni?