@hyperbolic2346 2022-06-15
> One more question: Why this is still here, not in rapids-spark-jni?

An excellent question. I plan on moving this to `spark-rapids-jni` shortly. I'm working on adding tests and benchmarks there now and once these two performance PRs are in place I will add a PR to move it out into that repo for good.